rename show hint component
