+++ client/modules/IDE/components/show-hint.js
